requirements.txt
Django==5.2.7
psycopg2-binary==2.9.9
channels==4.1.0
channels-redis==4.2.0
daphne==4.1.2
google-generativeai==0.8.3
requests==2.32.3
django-cors-headers==4.4.0
Pillow==10.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
python-decouple==3.8
whitenoise==6.7.0

